requirements.txt
aiofiles==0.8.0
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
astroid==2.9.3
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
beautifulsoup4==4.10.0
bleach==4.1.0
boto3==1.21.31
botocore==1.24.32
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
configparser2==4.0.0
cryptography==36.0.1
cssselect==1.1.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
dicttoxml==1.7.4
Django==4.1.1
django-allauth==0.51.0
djangorestframework==3.14.0
docutils==0.18.1
fake-useragent==0.1.11
filelock==3.8.0
fpdf==1.7.2
frozenlist==1.3.0
future==0.18.2
geocoder==1.38.1
geoip2==4.5.0
h11==0.13.0
idna==3.2
importlib-metadata==4.10.1
iniconfig==1.1.1
ip2geotools==0.1.6
IP2Location==8.6.4
isort==5.10.1
jmespath==1.0.0
keyring==23.5.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lxml==4.7.1
matplotlib==3.4.3
matrix==2.0.1
maxminddb==2.2.0
mccabe==0.6.1
multidict==6.0.2
nmap==0.0.1
numpy==1.21.4
oauthlib==3.2.2
outcome==1.1.0
packaging==21.3
Pillow==8.4.0
pip-review==1.1.1
pkginfo==1.8.2
platformdirs==2.4.1
pluggy==1.0.0
py==1.10.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
PyJWT==2.5.0
pylint==2.12.2
pynput==1.7.6
pyobjc-core==8.4.1
pyobjc-framework-ApplicationServices==8.4.1
pyobjc-framework-Cocoa==8.4.1
pyobjc-framework-Quartz==8.4.1
pyOpenSSL==22.0.0
pyparsing==3.0.4
pyquery==1.4.3
Pyrogram==2.0.73
PySocks==1.7.1
pyTelegramBotAPI==4.8.0
pytest==7.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
ratelim==0.1.6
readme-renderer==32.0
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
s3transfer==0.5.2
secret==0.8
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
snscrape==0.4.3.20220106
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tabulate==0.8.9
telebot==0.0.4
TelegramBotAPI==0.3.2
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
twine==3.8.0
typed-ast==1.5.2
types-cryptography==3.3.23.1
typing==3.7.4.3
typing_extensions==4.0.1
urllib3==1.26.7
webencodings==0.5.1
wrapt==1.13.3
wsproto==1.0.0
yarl==1.7.2
zipp==3.7.0
